+++ trunk/source/Library-Automata/src/de/uni_freiburg/informatik/ultimate/automata/nestedword/reachablestates/StateContainerFieldAndMap.java
@@ -1565,7 +1565,7 @@ class StateContainerFieldAndMap<LETTER, STATE> extends StateContainer<LETTER, ST
 					/**
 					 * Points to next field that has OutgoingInternalTransition.
 					 */
-					private short mPosition;
+					private int mPosition;
 					
 					{
 						mPosition = 0;
